Updated to Lewdos Version 1.2
Huge Updates. now focussing on actually creating an OS, rather than an assembler and Virtual assembly interpreter, because without proper User management, the assembler is hard to implement. implenemted a login system, that needs to be reworked, because its the most insecure login system ever. i also implementes the beginnings of "Memes" the memory manager and "Smash" the Bash like interfacing language for the OS.